requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
scikit-learn==1.2.2
numpy==1.24.3
thread6==1.0.0
joblib==1.2.0